% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{encryption}
\alias{encryption}
\title{ZENIT-POLAR substitution cipher method of encryption}
\usage{
encryption(x, key_from = "tenis", key_to = "polar")
}
\arguments{
\item{x}{message to encrypt}

\item{key_from}{first one-way cipher}

\item{key_to}{second one-way cipher}
}
\value{
encrypted message
}
\description{
Returns a message encrypted by ZENIT-POLAR substitution cipher method of encryption using by default TENIS-POLAR key.
}
\details{
This kind of encryption replace the nth character of 1th key to the nth character of 2nd key and vice versa.
}
\examples{

x <- "orro o um psibinhe tisi er kisir: e ivorre der ceseir, e celpsisae der cisopir!"
encryption(x, key_from = "polar", key_to = "tenis")

}
